# backend/gen/openapi/agent-platform.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "types/v1/types.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/runs": {
      "get": {
        "operationId": "ExecutionService_ListRuns",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListRunsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "team_id",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "agent_id",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "repository_binding_id",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "task_id",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "runtime",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "created_from",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "created_to",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "ExecutionService"
        ]
      }
    },
    "/v1/runs/{run_id}": {
      "get": {
        "operationId": "ExecutionService_GetRun",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1Run"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "ExecutionService"
        ]
      }
    },
...
    "/v1/runs/{run_id}/events": {
      "get": {
        "operationId": "RunEventStreamService_Stream",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "$ref": "#/definitions/v1StreamResponse"
                },
                "error": {
                  "$ref": "#/definitions/rpcStatus"
                }
              },
              "title": "Stream result of v1StreamResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "RunEventStreamService"
        ]
      }
    },
...
  },
  "definitions": {
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
...
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
...
    "v1Attempt": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "number": {
          "type": "integer",
          "format": "int32"
        },
        "worker_id": {
          "type": "string"
        },
        "state": {
          "type": "string"
        },
        "infrastructure_failure": {
          "type": "boolean"
        },
        "error": {
          "$ref": "#/definitions/v1ExecutionError"
        },
        "started_at": {
          "type": "string",
          "format": "date-time"
        },
        "ended_at": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
...
    "v1ExecutionError": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "v1ExecutionLimits": {
      "type": "object",
      "properties": {
        "timeout_seconds": {
          "type": "string",
          "format": "int64"
        },
        "cpus": {
          "type": "number",
          "format": "double"
        },
        "memory_bytes": {
          "type": "string",
          "format": "int64"
        },
        "pids": {
          "type": "string",
          "format": "int64"
        },
        "temp_bytes": {
          "type": "string",
          "format": "int64"
        },
        "egress": {
          "type": "string"
        }
      }
    },
...
    "v1ListRunsResponse": {
      "type": "object",
      "properties": {
        "items": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Run"
          }
        }
      }
    },
...
    "v1ModelBinding": {
      "type": "object",
      "properties": {
        "model_id": {
          "type": "string"
        }
      }
    },
    "v1ModelBudget": {
      "type": "object",
      "properties": {
        "max_input_tokens": {
          "type": "string",
          "format": "int64"
        },
        "max_output_tokens": {
          "type": "string",
          "format": "int64"
        },
        "max_cost_amount": {
          "type": "string"
        }
      }
    },
...
    "v1Run": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "session_id": {
          "type": "string"
        },
        "agent_release_id": {
          "type": "string"
        },
        "runtime_image_id": {
          "type": "string"
        },
        "request_text": {
          "type": "string"
        },
        "state": {
          "type": "string"
        },
        "model_binding": {
          "$ref": "#/definitions/v1ModelBinding"
        },
        "model_budget": {
          "$ref": "#/definitions/v1ModelBudget"
        },
        "execution_limits": {
          "$ref": "#/definitions/v1ExecutionLimits"
        },
        "usage": {
          "$ref": "#/definitions/v1Usage"
        },
        "cost_amount": {
          "type": "string"
        },
        "terminal_error": {
          "$ref": "#/definitions/v1ExecutionError"
        },
        "attempt_count": {
          "type": "integer",
          "format": "int32"
        },
        "created_by": {
          "type": "string"
        },
        "created_at": {
          "type": "string",
          "format": "date-time"
        },
        "started_at": {
          "type": "string",
          "format": "date-time"
        },
        "ended_at": {
          "type": "string",
          "format": "date-time"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time"
        },
        "version": {
          "type": "string",
          "format": "int64"
        },
        "attempts": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Attempt"
          }
        }
      }
    },
...
    "v1StreamResponse": {
      "type": "object",
      "properties": {
        "sequence": {
          "type": "string",
          "format": "int64"
        },
        "event_type": {
          "type": "string"
        },
        "payload": {},
        "created_at": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "v1Usage": {
      "type": "object",
      "properties": {
        "input_tokens": {
          "type": "string",
          "format": "int64"
        },
        "output_tokens": {
          "type": "string",
          "format": "int64"
        },
        "cost_amount": {
          "type": "string"
        }
      }
    },
...
  }
}